Add tests auth utils
