+++ packages/node-app/tsconfig.spec.json
@@ -9,6 +9,7 @@
     "jest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts",
+    "src/tests/__utils__/*.ts"
   ]
 }
